pyproject.toml
[project]
name = "elastic-agent-local"
version = "0.1.0"
description = "Local agentic RAG system using Elasticsearch, Ollama, and LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.0",
    "langchain-ollama>=0.3.0",
    "langchain-elasticsearch>=1.0.0",
    "langchain-community>=0.3.0",
    "langgraph>=1.0.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
    "elasticsearch>=8.17.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "duckduckgo-search>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
mcp = [
    "mcp>=1.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
